Move enqueuer package into internal/codeintel.
