+++ internal/codeintel/enqueuer/util.go
